+++ src/app/(provider)/(root)/plan/[planId]/page.tsx
@@ -14,7 +14,7 @@ import CreateScheduleButton from "../_components/CreateScheduleButton";
 import DayMenu from "../_components/DayMenu";
 import BarIcon from "../_components/icons/BarIcon";
 import MapIcon from "../_components/icons/MapIcon";
-import ScheculeList from "./ScheculeList";
+import ScheculeList from "./ScheduleList";
 
 const api = new PlanAPI(axios);
 type PlanDetailPageProps = { params: { planId: string } };
